requirements.txt
#pydantic<2.8
openai-whisper
#numpy<2
#numpy
#pyannote.audio
#onnxruntime
#torch
#torchvision
#git+https://github.com/m-bain/whisperX.git
telebot
python-dotenv
google-auth
google-api-python-client
langchain_openai
langchain_community
langchainhub
langchain-huggingface
gigachat
yandexcloud
#yandex_chain
langchain_google_genai
langchain_mistralai
llama-cpp-python

